+++ HISTORY.rst
@@ -142,7 +142,8 @@
   the solution.
 - Removing running workflow instances in document of a specific type if
   that document type is removed from the workflow.
-  
+- Make error messages persistent and increase the timeout of warning to 10 seconds.
+
 
 2.7.3 (2017-09-11)
 ==================